Updated install routine to properly trust CA root on Ubuntu Desktop

+++ commands/install.cmd
@@ -23,18 +23,34 @@ if [[ ! -f "${WARDEN_SSL_DIR}/rootca/certs/ca.cert.pem" ]]; then
     -subj "/C=US/O=Warden Proxy Local CA"
 fi
 
-## trust root ca differently on linux-gnu than on macOS
-if [[ "$OSTYPE" == "linux-gnu" ]] && [[ ! -f /etc/pki/ca-trust/source/anchors/warden-proxy-local-ca.cert.pem ]]; then
-  echo "==> Trusting root certificate (requires sudo privileges)"
+## trust root ca differently on Fedora, Ubuntu and macOS
+if [[ "$OSTYPE" == "linux-gnu" ]] \
+  && [[ -d /etc/pki/ca-trust/source/anchors ]] \
+  && [[ ! -f /etc/pki/ca-trust/source/anchors/warden-proxy-local-ca.cert.pem ]] \
+  ## Fedora/CentOS
+then
+  echo "==> Trusting root certificate (requires sudo privileges)"  
   sudo cp "${WARDEN_SSL_DIR}/rootca/certs/ca.cert.pem" /etc/pki/ca-trust/source/anchors/warden-proxy-local-ca.cert.pem
   sudo update-ca-trust
   sudo update-ca-trust enable
-elif [[ "$OSTYPE" == "darwin"* ]] && ! security dump-trust-settings -d | grep 'Warden Proxy Local CA' >/dev/null; then
+elif [[ "$OSTYPE" == "linux-gnu" ]] \
+  && [[ -d /usr/local/share/ca-certificates ]] \
+  && [[ ! -f /usr/local/share/ca-certificates/warden-proxy-local-ca.cert.pem ]] \
+  ## Ubuntu/Debian
+then
+  echo "==> Trusting root certificate (requires sudo privileges)"  
+  sudo cp "${WARDEN_SSL_DIR}/rootca/certs/ca.cert.pem" /usr/local/share/ca-certificates/warden-proxy-local-ca.cert.pem
+  sudo update-ca-certificates
+elif [[ "$OSTYPE" == "darwin"* ]] \
+  && ! security dump-trust-settings -d | grep 'Warden Proxy Local CA' >/dev/null \
+  ## Apple macOS
+then
   echo "==> Trusting root certificate (requires sudo privileges)"
   sudo security add-trusted-cert -d -r trustRoot \
     -k /Library/Keychains/System.keychain "${WARDEN_SSL_DIR}/rootca/certs/ca.cert.pem"
 fi
 
+## sign certificate used by services run on warden.test sub-domains
 if [[ ! -f "${WARDEN_SSL_DIR}/certs/warden.test.crt.pem" ]]; then
   "${WARDEN_DIR}/bin/warden" sign-certificate warden.test
 fi
